+++ packages/xprog-prensa/news/PageBlock/component.tsx
@@ -23,10 +23,13 @@ export const PageBlock: React.FC<PageBlockProps> = ({
   slotManualExtra,
   slotManualExtraTop,
   slotLeftBgColor,
+  slotLeftColor,
   slotLeftLayout,
   slotCenterBgColor,
+  slotCenterColor,
   slotCenterLayout,
   slotRightBgColor,
+  slotRightColor,
   slotRightLayout,
   slotLayouts,
   slotTemplate,
@@ -38,7 +41,10 @@ export const PageBlock: React.FC<PageBlockProps> = ({
     const layoutSlotConfig = get(layoutSelected, `slotConfig[${slotTemplate}][slot${name}]`)
     return {
       css: layoutSelected?.css,
-      ...layoutSlotConfig
+      ...layoutSlotConfig,
+      wrapperTop: layoutSelected?.wrapperTop,
+      wrapperContent: layoutSelected?.wrapperContent,
+      wrapperBottom: layoutSelected?.wrapperBottom
     }
   }
   const leftColumnLayout = getLayoutBySlotAndTemplate('Left', slotLeftLayout)
@@ -125,11 +131,15 @@ export const PageBlock: React.FC<PageBlockProps> = ({
         ...template?.css?.item
       }
     },
+    color: slotLeftColor,
     items: slotAutoLeftElements,
     itemsBottom: slotManualBottomLeft,
     itemsTop: slotManualTopLeft,
     layout: leftColumnLayout,
-    name: 'Left'
+    name: 'Left',
+    wrapperTop: leftColumnLayout.wrapperTop,
+    wrapperContent: leftColumnLayout.wrapperContent,
+    wrapperBottom: leftColumnLayout.wrapperBottom
   }
   const columnCenterProps: ColumnProps = {
     ...columnProps,
@@ -144,11 +154,15 @@ export const PageBlock: React.FC<PageBlockProps> = ({
         ...template?.css?.item
       }
     },
+    color: slotCenterColor,
     items: slotAutoCenterElements,
     itemsBottom: slotManualBottomCenter,
     itemsTop: slotManualTopCenter,
     layout: centerColumnLayout,
-    name: 'Center'
+    name: 'Center',
+    wrapperTop: centerColumnLayout.wrapperTop,
+    wrapperContent: centerColumnLayout.wrapperContent,
+    wrapperBottom: centerColumnLayout.wrapperBottom
   }
   const columnRightProps: ColumnProps = {
     ...columnProps,
@@ -163,11 +177,15 @@ export const PageBlock: React.FC<PageBlockProps> = ({
         ...template?.css?.item
       }
     },
+    color: slotRightColor,
     items: slotAutoRightElements,
     itemsBottom: slotManualBottomRight,
     itemsTop: slotManualTopRight,
     layout: rightColumnLayout,
-    name: 'Right'
+    name: 'Right',
+    wrapperTop: rightColumnLayout.wrapperTop,
+    wrapperContent: rightColumnLayout.wrapperContent,
+    wrapperBottom: rightColumnLayout.wrapperBottom
   }
   const columnExtraProps: ColumnProps = {
     ...columnProps,