+++ packages/xprog-prensa/primitives/Accordion/ItemGroup.tsx
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import { useLink } from '../../hooks/useLink'
 import { Block } from '../Block'
 import { IconButton } from '../IconButton'
+import { Link } from '../Link'
 import { Typography } from '../Typography'
 import { ExpandLessIcon } from './ExpandLessIcon'
 import { ExpandMoreIcon } from './ExpandMoreIcon'
@@ -15,6 +15,8 @@ export const AccordionItemGroup: React.FC<AccordionItemGroupProps> = ({
   accordionTitleAs,
   accordionTitleVariant,
   children,
+  customLink,
+  customLinkProps,
   css,
   expanded,
   href,
@@ -23,7 +25,6 @@ export const AccordionItemGroup: React.FC<AccordionItemGroupProps> = ({
   iconNotExpandedState,
   id,
   innerSpace,
-  linkProps,
   onClick,
   title
 }) => {
@@ -98,8 +99,20 @@ export const AccordionItemGroup: React.FC<AccordionItemGroupProps> = ({
         variant={accordionTitleVariant}
         css={title_css}
       >
-        {useLink({ bypass: !href ? true : false, href, ...linkProps }, title)}
-        {useLink({ bypass: children ? true : false, href, ...linkProps }, (
+        <Link
+          bypass={!href}
+          component={customLink}
+          href={href}
+          {...customLinkProps}
+        >
+          {title}
+        </Link>
+        <Link
+          bypass={children ? true : false}
+          component={customLink}
+          href={href}
+          {...customLinkProps}
+        >
           <IconButton
             className='pds-Accordion-ItemGroup-IconButton'
             variant='ghost'
@@ -107,7 +120,7 @@ export const AccordionItemGroup: React.FC<AccordionItemGroupProps> = ({
             icon={itemIcon}
             css={{ button: icon_button_css, icon: icon_css }}
           />
-        ))}
+        </Link>
       </Typography>
       <AccordionContent
         className='pds-Accordion-ItemGroup-content'