+++ packages/xprog-prensa/primitives/Accordion/component.tsx
@@ -7,10 +7,12 @@ import { AccordionProps, AccordionItemGroupProps } from './types'
 
 export const Accordion: React.FC<AccordionProps> = ({
   customLink,
-  customLinkProps,
   css,
   items,
   innerSpace,
+  icon,
+  iconExpandedState,
+  iconNotExpandedState,
   size,
 }) => {
   const [state, setState] = useState(items || [])
@@ -49,11 +51,13 @@ export const Accordion: React.FC<AccordionProps> = ({
       {state.map(({ id, onClick, ...otherProps }: AccordionItemGroupProps) => (
         <AccordionItemGroup
           customLink={customLink}
-          customLinkProps={customLinkProps}
           css={accordionitem_css}
           key={id}
           id={id}
           innerSpace={innerSpace}
+          icon={otherProps.icon ? otherProps.icon : icon}
+          iconExpandedState={otherProps.iconExpandedState ? otherProps.iconExpandedState : iconExpandedState}
+          iconNotExpandedState={otherProps.iconNotExpandedState ? otherProps.iconNotExpandedState : iconNotExpandedState}
           onClick={onClick ? onClick : () => handleClick(id)}
           {...otherProps}
         />