+++ packages/xprog-prensa/primitives/Calendar/types.ts
@@ -1,0 +1,19 @@
+export interface CalendarProps {
+  onChange: (P: Date) => void;
+  children?: any;
+  css?: {
+    container?: {};
+    headerGroup?: {};
+    headerButton?: {};
+    headerTitle?: {};
+    weekGroup?: {};
+    weekDay?: {};
+    dayGroup?: {};
+    day?: {};
+    dayButton?: {};
+  };
+  date?: Date;
+  icBackward?: any;
+  icForward?: any;
+  selectedDate?: Date;
+}