+++ packages/xprog-prensa/primitives/Link/styled.tsx
@@ -1,0 +1,5 @@
+import { PrensaEngine } from '../../prensa.default.config'
+
+export const StyledA = PrensaEngine.styled('a', {})
+
+export type StyledAProps = typeof StyledA